Reshuffling of classes, local imports, more signals
Now all the QFrames are defined in `noteorganiser/frames.py`, and the
only definition in `noteorganiser/widgets.py` is currently the
`PicButton`.

All the files except the main script have now relative imports -- hope
that fixes everything.

Finally, all wiggly calls to a parent's parent function have now been
replaced by signals and slots. Should improve the convenience of
testing.

The incriminating part of the test (checking the presence of a signal
throught the qtbot) has been commented out, waiting for a solution to
the problem.