add toolbar icons
we use qtawesome as the icon pack.
This uses qtpy internally but we can't move to qtpy completely at the
moment, because there's an issue with the QWebView. So we have to set
the environment variable 'QT_API' to 'PySide' that qtawesome plays
nicely together with our other QT elements